requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
django-unfold==0.23.0
djangorestframework==3.14.0
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-ratelimit==4.1.0
requests==2.31.0
